test-strapi-c8-hook/requirement.txt
argparse
logging
prettytable
pynpm
requests
GitPython
